Clean up - rename files, move commands to own file, etc.
