+++ pages/index.ts
@@ -15,7 +15,7 @@ import "../src/css/sidebar.scss";
 import "../src/css/validationControls.scss";
 import "../src/css/validationSidebarOutput.scss";
 import createValidatorPlugin from "../src/ts/index";
-import createView from "../src/ts/view";
+import createView from "../src/ts/createView";
 import regexAdapter from "../src/ts/adapters/regex";
 
 const mySchema = new Schema({
